+++ CHANGELOG.md
@@ -44,7 +44,7 @@ This project tries to follow [SemVer 2.0.0](https://semver.org/).
     wharf-cmd-workers from the wharf-cmd-provisioner and kills them in an effort
     to clean up forgotten builds/workers.
 
-- Added new implementation for `wharf run`. (#33, #45)
+- Added new implementation for `wharf run`. (#33, #45, #66)
 
 - Added new implementation for `.wharf-ci.yml` file parsing that now supports
   returning multiple errors for the whole parsing as well as keep track of the